Share Unix socket deadline helpers
Centralize scoped timeout restoration and absolute deadline refresh for framed and shared-memory Unix transport I/O.


Copilot-Session: ac86dd2d-0280-4d7b-87a8-5654ad1503a6
